Fix system-info support for Unicode characters on Windows (#7626)
- The computer_name and local_hostname columns where incorrectly
  truncating Unicode characters when the computer name contained them.
  Use the WinAPI to get the computer name instead of the POSIX one,
  which correctly supports Unicode.

- Minor cleanup on the getHostname implementation of Linux and macOS too,
  since POSIX guarantees the same max size.
