+++ osquery/utils/system/posix/system.h
@@ -13,8 +13,8 @@
 #include <memory>
 #include <string>
 
-#include <boost/filesystem/path.hpp>
 #include <boost/core/noncopyable.hpp>
+#include <boost/filesystem/path.hpp>
 #include <gtest/gtest_prod.h>
 #include <osquery/utils/info/tool_type.h>
